feat: 🎸 support add_torrent from torrent file
change the add_torrent method and AddTorrentArg and it's source to
support bytes from file.

BREAKING CHANGE: 🧨 The add_torrent method no longer uses the generic post and request
methods, but given that there is only one place in the qbittorrent api
that sends raw data, such a hack is acceptable.

✅ Closes: #11